Update structure for generated config.c #chonkerkeys/10
